fix(bridge): restore Unity 6.0-6.4 compilation broken by 6000.5 EntityId compat
PR #2's EntityId work widened GetSceneHandle's return type to long and, in the
pre-6000.5 (#else) branch, returned Scene.handle directly. On Unity 6000.0-6000.4
Scene.handle is a SceneHandle with implicit int and uint operators, so widening
straight to long is ambiguous (CS0457) and the entire editor bridge failed to
compile across the 6.0-6.4 support matrix. Pin the int conversion explicitly --
handles are 32-bit on these versions, so it is lossless.

Validated by running the editmode suite on 6000.4.0f1 (compiles clean, tests pass).

+++ unity-package/com.ucp.bridge/Editor/Compatibility/UnityObjectCompat.cs
@@ -20,7 +20,10 @@ namespace UCP.Bridge
 #if UNITY_6000_5_OR_NEWER
             return unchecked((long)scene.handle.GetRawData());
 #else
-            return scene.handle;
+            // On 6000.0–6000.4, Scene.handle is a SceneHandle with implicit int and
+            // uint operators; widening straight to long is ambiguous (CS0457). Pin the
+            // int conversion explicitly — handles are 32-bit on these versions.
+            return (int)scene.handle;
 #endif
         }
